refactor: edit types
