+++ async-race/src/modules/model/model-i.ts
@@ -1,0 +1,1 @@
+export type ISwitch = { [key: string]: boolean };
