+++ async-race/src/modules/state/state-i.ts
@@ -1,0 +1,12 @@
+import { ICar } from '../types.ts/types';
+
+export interface ISavedState {
+  pageNum: number;
+  cars: Array<ICar | undefined>;
+  totalCars: number;
+  page: string;
+  id: string;
+  animation: { [key: string]: number };
+  race: boolean;
+  winnerFound: boolean;
+}
