+++ async-race/src/modules/types.ts/types.ts
@@ -4,25 +4,8 @@ export enum Path {
   Winners = '/winners',
 }
 
-// export type TButtons = { [key: string]: HTMLButtonElement | null };
-
-// export type TDisabled = { [key: string]: HTMLElement | null };
-
 export interface ICar {
   name: string;
   color: string;
   id: number;
 }
-
-export interface ISavedState {
-  pageNum: number;
-  cars: Array<ICar | undefined>;
-  totalCars: number;
-  page: string;
-  id: string;
-  animation: { [key: string]: number };
-  race: boolean;
-  winnerFound: boolean;
-}
-
-export type ISwitch = { [key: string]: boolean };
